+++ src/addons/display_commander/nvapi/nvapi_loader.hpp
@@ -29,6 +29,10 @@ using NvAPI_DISP_GetAdaptiveSyncData_pfn = NvAPI_Status(__cdecl*)(NvU32 displayI
 using NvAPI_DISP_GetDisplayIdByDisplayName_pfn = NvAPI_Status(__cdecl*)(const char* displayName, NvU32* displayId);
 using NvAPI_Disp_GetVRRInfo_pfn = NvAPI_Status(__cdecl*)(NvU32 displayId, NV_GET_VRR_INFO* pVrrInfo);
 
+using NvAPI_EnumPhysicalGPUs_pfn = NvAPI_Status(__cdecl*)(NvPhysicalGpuHandle* nvGPUHandle, NvU32* pGpuCount);
+using NvAPI_GPU_GetDynamicPstatesInfoEx_pfn =
+    NvAPI_Status(__cdecl*)(NvPhysicalGpuHandle hPhysicalGpu, NV_GPU_DYNAMIC_PSTATES_INFO_EX* pDynamicPstatesInfoEx);
+
 using NvAPI_DRS_CreateSession_pfn = NvAPI_Status(__cdecl*)(NvDRSSessionHandle* phSession);
 using NvAPI_DRS_DestroySession_pfn = NvAPI_Status(__cdecl*)(NvDRSSessionHandle hSession);
 using NvAPI_DRS_LoadSettings_pfn = NvAPI_Status(__cdecl*)(NvDRSSessionHandle hSession);
@@ -74,6 +78,8 @@ struct NvApiPtrs {
     NvAPI_DISP_GetAdaptiveSyncData_pfn DISP_GetAdaptiveSyncData = nullptr;
     NvAPI_DISP_GetDisplayIdByDisplayName_pfn DISP_GetDisplayIdByDisplayName = nullptr;
     NvAPI_Disp_GetVRRInfo_pfn Disp_GetVRRInfo = nullptr;
+    NvAPI_EnumPhysicalGPUs_pfn EnumPhysicalGPUs = nullptr;
+    NvAPI_GPU_GetDynamicPstatesInfoEx_pfn GPU_GetDynamicPstatesInfoEx = nullptr;
 
     NvAPI_DRS_CreateSession_pfn DRS_CreateSession = nullptr;
     NvAPI_DRS_DestroySession_pfn DRS_DestroySession = nullptr;